feat(mmrelay.meshtastic.metadata): add metadata utilities and export helpers
Move metadata-related helpers from meshtastic_utils.py into a new
metadata module, exposing a stable set of internal APIs for firmware
and name extraction. Update imports to consume the new APIs and remove
duplicated implementations from the old module. Expose check_connection
via the health module to enable centralized connection health checks.
